refactor(library): extract shared useFreeformEndpointNode hook
`getNodeRect`, `findFreeformEndpointNode`, and the `FREEFORM_ENDPOINT_SNAP_RADIUS_PX`
constant were byte-identical in both edge path hooks. Move them into one
`useFreeformEndpointNode` hook that both consume, removing the duplication (and
its now-unused Node/isParentNodeType/distanceToRect imports).

Verified: tsc, lint, and 44 edge e2e (freeform-anchor, edge-interactions,
shape-aware, ghost-preview, fresh-bend, handle-zoom).
